refactor(services): separate services
Separate services with the types
