+++ src/store/authStore.ts
@@ -1,0 +1,13 @@
+import { create } from 'zustand';
+
+export interface AuthState {
+  accessToken: string | null;
+  setAccessToken: (token: string | null) => void;
+  logout: () => void;
+}
+
+export const useAuthStore = create<AuthState>((set) => ({
+  accessToken: null,
+  setAccessToken: (token) => set({ accessToken: token }),
+  logout: () => set({ accessToken: null }),
+}));
