+++ src/types/homes-types.ts
@@ -1,0 +1,8 @@
+export interface CreateHomeData {
+  homeName: string;
+}
+export interface Home {
+  _id: string;
+  homeName: string;
+  rooms: string[];
+}
