+++ src/types/room-types.ts
@@ -1,0 +1,9 @@
+export interface CreateRoomData {
+  roomName: string;
+}
+export interface Room {
+  _id: string;
+  roomName: string;
+  homeId: string;
+  devices: string[];
+}
